feat(plugins): extract generic v1/models fetching
