+++ plugins/cmd/litellm/main.go
@@ -10,6 +10,7 @@ import (
 	"strings"
 	"time"
 
+	"github.com/naira-project/naira/plugins/internal/openaicompat"
 	"github.com/naira-project/naira/plugins/pkg/pluginapi"
 	"github.com/naira-project/naira/plugins/pkg/pluginmain"
 	"k8s.io/client-go/dynamic"
@@ -52,14 +53,12 @@ func New(config config, logger *log.Logger) *Plugin {
 
 func main() {
 	app := pluginmain.New[config]()
-
 	p := New(app.PluginConfig, app.Logger)
-
 	app.Serve(p)
 }
 
 func (p *Plugin) Collect(ctx context.Context) (pluginapi.CollectResponse, error) {
-	models, err := p.fetchModels(ctx)
+	models, err := openaicompat.FetchModels(ctx, p.httpClient, p.config.BaseURL, p.config.APIKey)
 	if err != nil {
 		return pluginapi.CollectResponse{}, fmt.Errorf("fetching LiteLLM models: %w", err)
 	}
@@ -156,31 +155,6 @@ func (p *Plugin) Collect(ctx context.Context) (pluginapi.CollectResponse, error)
 	return response, nil
 }
 
-func (p *Plugin) fetchModels(ctx context.Context) ([]model, error) {
-	req, err := http.NewRequestWithContext(ctx, http.MethodGet, p.config.BaseURL+"/v1/models", nil)
-	if err != nil {
-		return nil, fmt.Errorf("building LiteLLM models request: %w", err)
-	}
-	p.addAuthorization(req)
-
-	resp, err := p.httpClient.Do(req)
-	if err != nil {
-		return nil, fmt.Errorf("calling LiteLLM models endpoint: %w", err)
-	}
-	defer resp.Body.Close()
-
-	if resp.StatusCode < 200 || resp.StatusCode >= 300 {
-		return nil, fmt.Errorf("litellm /v1/models returned %s", resp.Status)
-	}
-
-	var payload modelsResponse
-	if err := json.NewDecoder(resp.Body).Decode(&payload); err != nil {
-		return nil, fmt.Errorf("decoding LiteLLM models response: %w", err)
-	}
-
-	return payload.Data, nil
-}
-
 func (p *Plugin) fetchAllowedModels(ctx context.Context, key string) ([]string, error) {
 	req, err := http.NewRequestWithContext(ctx, http.MethodGet, p.config.BaseURL+"/key/info", nil)
 	if err != nil {
@@ -190,7 +164,10 @@ func (p *Plugin) fetchAllowedModels(ctx context.Context, key string) ([]string, 
 	query := req.URL.Query()
 	query.Set("key", key)
 	req.URL.RawQuery = query.Encode()
-	p.addAuthorization(req)
+
+	if strings.TrimSpace(p.config.APIKey) != "" {
+		req.Header.Set("Authorization", "Bearer "+p.config.APIKey)
+	}
 
 	resp, err := p.httpClient.Do(req)
 	if err != nil {
@@ -208,12 +185,6 @@ func (p *Plugin) fetchAllowedModels(ctx context.Context, key string) ([]string, 
 	}
 
 	return payload.Info.Models, nil
-}
-
-func (p *Plugin) addAuthorization(req *http.Request) {
-	if strings.TrimSpace(p.config.APIKey) != "" {
-		req.Header.Set("Authorization", "Bearer "+p.config.APIKey)
-	}
 }
 
 func newAppIdentityProvider(logger *log.Logger) AppIdentityProvider {
@@ -252,15 +223,6 @@ func dedupeNodes(nodes []pluginapi.NodeClaim) []pluginapi.NodeClaim {
 	}
 
 	return result
-}
-
-type modelsResponse struct {
-	Data []model `json:"data"`
-}
-
-type model struct {
-	ID      string `json:"id"`
-	OwnedBy string `json:"owned_by"`
 }
 
 type keyInfoResponse struct {
